feat: changed to using JWT tokens for authorization instead of just passing user.id

+++ client/app/api/flashcards/delete/route.js
@@ -1,12 +1,19 @@
 import { db } from "@/firebase";
+import { auth } from "@clerk/nextjs/server";
 import { deleteDoc, doc } from "firebase/firestore";
 import { NextResponse } from "next/server";
 
 export async function POST(req) {
-    const data = await req.json();
-    const { userId, flashcardId } = data;
+    const { userId } = auth();
+
+    if (!userId) {
+        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
 
     try {
+        const data = await req.json();
+        const { flashcardId } = data;
+        
         await deleteDoc(doc(db, "users", userId, "flashcardSets", flashcardId));
         return NextResponse.json({ message: "Successfully deleted" }, { status: 200 });
     } catch (error) {
